+++ core/mesh.h
@@ -16,9 +16,9 @@ typedef struct{
 } Mesh;
 
 Triangle makeTriangle(unsigned int a, unsigned int b, unsigned int c);
-void addVertexToMesh(Mesh *mesh, float x, float y, float z);
+void addVertexToMesh(Mesh *mesh, double x, double y, double z);
 void addFaceToMesh(Mesh *mesh, unsigned int a, unsigned int b, unsigned int c);
-Mesh *createMeshFromHeightfield(float **heightfield, const int size);
+Mesh *createMeshFromHeightfield(double **heightfield, const int size);
 void exportMeshAsObj(Mesh *mesh, const char *filename);
 void freeMesh(Mesh *mesh);
