+++ core/noise.cpp
@@ -56,6 +56,6 @@ void loadNoisePermutation(char* fileName){
 	for (int i=0; i < 256 ; i++) p[256+i] = p[i] = permutation[i];
 }
 
-float randRange(float min, float max){
-    return min + (max - min) * ((float)rand() / RAND_MAX);
+double randRange(double min, double max){
+    return min + (max - min) * ((double)rand() / RAND_MAX);
 }
